SPA refactor + full bug audit fixes
Frontend — SPA architecture:
- All views extracted to components/views/ (ViewRouter pattern)
- app/*/page.tsx stripped to redirect stubs (setView + router.replace)
- AppShell transitions keyed on currentView instead of pathname
- Sidebar uses setView() buttons instead of <Link> for zero-latency nav
- DataPrefetcher fires all queries on mount with exact matching cache keys
- Hover prefetch keys now match view query keys exactly (was silently missing cache)
- Fixed emi-forecast cache collision: ReportsView uses ['emi-forecast', 12]

Frontend — Decimal-as-string bugs:
- DashboardView: Number() wrap on utilization_pct, monthly_emi_burden, total_receivables comparisons
- DashboardView: Math.round(Number()) for card utilization_pct display
- ReportsView: same utilization_pct comparison + display fixes
- SpendingChart: normalize MonthlyTrend Decimal fields to Number before passing to Recharts
- CategoryChart: normalize amount/percentage to Number; Math.round for percentage display

Backend — data integrity:
- emis.py: _refresh_friend_totals() now called on create/update/delete/record-payment for FRIEND EMIs (totals were never syncing)
- emis.py: EMI forecast month comparison changed to year-month ordinal to avoid end-of-month edge cases
- models/friend.py: active_emi_count Integer instead of Numeric(4,0)
- schemas/friend.py: active_emi_count: int
- alembic 0002: migration to fix active_emi_count column type

+++ backend/app/models/friend.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Boolean, DateTime, Numeric, ForeignKey, Text
+from sqlalchemy import Column, String, Boolean, DateTime, Numeric, Integer, ForeignKey, Text
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
@@ -26,7 +26,7 @@ class Friend(Base):
     total_emi_amount = Column(Numeric(12, 2), default=0)
     total_collected = Column(Numeric(12, 2), default=0)
     total_pending = Column(Numeric(12, 2), default=0)
-    active_emi_count = Column(Numeric(4, 0), default=0)
+    active_emi_count = Column(Integer, default=0)
     risk_level = Column(String(10), default="LOW")  # LOW, MEDIUM, HIGH
 
     created_at = Column(DateTime(timezone=True), server_default=func.now())
